test/e2e/pyproject.toml
[tool.poetry]
name = "pytest-bdd-rest"
version = "0.1.0"
description = ""
authors = ["Gaëtan Lehmann <gaetan.lehmann@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pytest = "^7.1.2"
requests = "^2.28.1"
pytest-playwright = "^0.3.0"
pdbpp = "^0.10.3"
pytest-xdist = "^2.5.0"
pytest-html = "^3.1.1"
ansi2html = "^1.8.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
